Stub out extendable tests + ModelManifest & ExtendableTestCase
